Switch to pydata-sphinx-theme.
The plan is to link to the notebooks repo directly from the
main docs, so it makes sense that the themes be consistent.

Although the sphinx-book-theme has some nice features and is
pretty similar, it would be nice to have the themes between
the main docs and notebooks be identical.